+++ bpl/extended_dixon_coles.py
@@ -45,8 +45,8 @@ class ExtendedDixonColesMatchPredictor(BaseMatchPredictor):
         home_team: jnp.array,
         away_team: jnp.array,
         num_teams: int,
-        home_goals: Iterable[float],
-        away_goals: Iterable[float],
+        home_goals: Iterable[int],
+        away_goals: Iterable[int],
         team_covariates: Optional[np.array],
     ):
         std_home_advantage = numpyro.sample(
@@ -218,8 +218,8 @@ class ExtendedDixonColesMatchPredictor(BaseMatchPredictor):
         self,
         home_team: Union[str, Iterable[str]],
         away_team: Union[str, Iterable[str]],
-        home_goals: Union[float, Iterable[float]],
-        away_goals: Union[float, Iterable[float]],
+        home_goals: Union[int, Iterable[int]],
+        away_goals: Union[int, Iterable[int]],
     ) -> jnp.array:
 
         home_team = [home_team] if isinstance(home_team, str) else home_team
